requirements.txt

pymssql
pandas
names
python-dotenv